Consolidate Project Models
Consolidates project models into a single file to avoid circular imports and allow
project models to reference each other in type annotations.
